added input params to weekly uploader lambda to be used in weekly scheduled execution
